requirements.txt
python-dotenv
langchain
langchain_openai
youtube_transcript_api
gradio